Finish Lab4

+++ Lab4/main.c
@@ -5,28 +5,41 @@
 
 int parr[2]= {0, 0};
 
-void sigchld_handler(int signal) {
+void signal_handler(int sig) {
   int status, pid;
 
   while ((pid = wait(&status)) != -1) {
     if(WIFEXITED(status)) {
       printf("\nA child with pid %d terminated with exit code %d\n", pid, WEXITSTATUS(status));
-      if (pid == parr[0]) {
-        if (WEXITSTATUS(status) == 1) {
-          kill(parr[1], SIGKILL);
-        } else {
-          kill(parr[1], SIGCONT);
+      if (sig == SIGCHLD) {
+        if (pid == parr[0]) {
+          if (WEXITSTATUS(status) == 1) {
+            kill(parr[1], SIGKILL);
+          } else {
+            kill(parr[1], SIGCONT);
+          }
+        } else if (pid == parr[1]) {
+          raise(SIGTERM);
         }
-      } else if (pid == parr[1]) {
-        raise(SIGTERM);
       }
     }
   }
+
+  if (sig == SIGCONT) {
+   printf("cont");
+  } else if (sig == SIGKILL) {
+   printf("kill");
+  }
+
+  signal(sig, signal_handler);
 }
 
 int main()
 {
-    signal(SIGCHLD, sigchld_handler);  
+    signal(SIGCHLD, signal_handler);  
+    signal(SIGCONT, signal_handler);
+    signal(SIGKILL, signal_handler);
+
 
     int pid, X;
